renamed connectInsights to connectHitsInsights
