+++ packages/react-instantsearch-core/src/connectors/connectHitsInsights.ts
