+++ packages/react-instantsearch-core/src/index.js
@@ -48,4 +48,6 @@ export { default as connectStats } from './connectors/connectStats';
 export {
   default as connectToggleRefinement,
 } from './connectors/connectToggleRefinement';
-export { default as connectInsights } from './connectors/connectInsights';
+export {
+  default as connectHitsInsights,
+} from './connectors/connectHitsInsights';
