+++ stories/InfiniteHits.stories.js
@@ -10,7 +10,7 @@ import {
 } from 'react-instantsearch-dom';
 import { WrapWithHits } from './util';
 import { action } from '@storybook/addon-actions';
-import { connectInsights } from 'react-instantsearch-core';
+import { connectHitsInsights } from 'react-instantsearch-core';
 
 const stories = storiesOf('InfiniteHits', module);
 
@@ -55,7 +55,7 @@ stories
   .add('with Insights', () => {
     const insightsClient = (method, payload) =>
       action(`[InsightsClient] sent ${method} with payload`)(payload);
-    const ProductWithInsights = connectInsights(insightsClient)(Product);
+    const ProductWithInsights = connectHitsInsights(insightsClient)(Product);
 
     function Product({ hit, insights }) {
       return (
